+++ src/core-interfaces.ts
@@ -213,19 +213,7 @@ export interface DataOrPathProps {
 	 */
 	data?: string
 }
-export interface BackgroundProps extends DataOrPathProps, ShapeFillProps {
-	/**
-	 * Color (hex format)
-	 * @deprecated v3.6.0 - use `ShapeFillProps` instead
-	 */
-	fill?: HexColor
-
-	/**
-	 * source URL
-	 * @deprecated v3.6.0 - use `DataOrPathProps` instead - remove in v4.0.0
-	 */
-	src?: string
-}
+export interface BackgroundProps extends DataOrPathProps, ShapeFillProps {}
 /**
  * Color in Hex format
  * @example 'FF3399'
@@ -427,12 +415,6 @@ export interface ShapeFillProps {
 	 * @default 'solid'
 	 */
 	type?: 'none' | 'solid'
-
-	/**
-	 * Transparency (percent)
-	 * @deprecated v3.3.0 - use `transparency`
-	 */
-	alpha?: number
 }
 /**
  * A single colour stop within a gradient fill.
@@ -533,30 +515,6 @@ export interface ShapeLineProps extends ShapeFillProps {
 	endArrowType?: 'none' | 'arrow' | 'diamond' | 'oval' | 'stealth' | 'triangle'
 	// FUTURE: beginArrowSize (1-9)
 	// FUTURE: endArrowSize (1-9)
-
-	/**
-	 * Dash type
-	 * @deprecated v3.3.0 - use `dashType`
-	 */
-	lineDash?: 'solid' | 'dash' | 'dashDot' | 'lgDash' | 'lgDashDot' | 'lgDashDotDot' | 'sysDash' | 'sysDot'
-	/**
-	 * @deprecated v3.3.0 - use `beginArrowType`
-	 */
-	lineHead?: 'none' | 'arrow' | 'diamond' | 'oval' | 'stealth' | 'triangle'
-	/**
-	 * @deprecated v3.3.0 - use `endArrowType`
-	 */
-	lineTail?: 'none' | 'arrow' | 'diamond' | 'oval' | 'stealth' | 'triangle'
-	/**
-	 * Line width (pt)
-	 * @deprecated v3.3.0 - use `width`
-	 */
-	pt?: number
-	/**
-	 * Line size (pt)
-	 * @deprecated v3.3.0 - use `width`
-	 */
-	size?: number
 }
 // used by: chart, slide, table, text
 export interface TextBaseProps {
@@ -631,30 +589,6 @@ export interface TextBaseProps {
 		 * @example 10 // numbered bullets start with 10
 		 */
 		numberStartAt?: number
-
-		// DEPRECATED
-
-		/**
-		 * Bullet code (unicode)
-		 * @deprecated v3.3.0 - use `characterCode`
-		 */
-		code?: string
-		/**
-		 * Margin between bullet and text
-		 * @since v3.2.1
-		 * @deplrecated v3.3.0 - use `indent`
-		 */
-		marginPt?: number
-		/**
-		 * Number to start with (only applies to type:number)
-		 * @deprecated v3.3.0 - use `numberStartAt`
-		 */
-		startAt?: number
-		/**
-		 * Number type
-		 * @deprecated v3.3.0 - use `numberType`
-		 */
-		style?: string
 	}
 	/**
 	 * Text color
@@ -1067,28 +1001,6 @@ export interface ShapeProps extends PositionProps, ObjectNameProps {
 	 * @since v4.2.0
 	 */
 	bevel?: Shape3DProps
-
-	/**
-	 * @deprecated v3.3.0
-	 */
-	lineSize?: number
-	/**
-	 * @deprecated v3.3.0
-	 */
-	lineDash?: 'dash' | 'dashDot' | 'lgDash' | 'lgDashDot' | 'lgDashDotDot' | 'solid' | 'sysDash' | 'sysDot'
-	/**
-	 * @deprecated v3.3.0
-	 */
-	lineHead?: 'arrow' | 'diamond' | 'none' | 'oval' | 'stealth' | 'triangle'
-	/**
-	 * @deprecated v3.3.0
-	 */
-	lineTail?: 'arrow' | 'diamond' | 'none' | 'oval' | 'stealth' | 'triangle'
-	/**
-	 * Shape name (used instead of default "Shape N" name)
-	 * @deprecated v3.10.0 - use `objectName`
-	 */
-	shapeName?: string
 }
 
 /**
@@ -1434,15 +1346,6 @@ export interface TableToSlidesProps extends TableProps {
 	 * - this margin will be across all slides created by auto-paging
 	 */
 	slideMargin?: Margin
-
-	/**
-	 * @deprecated v3.3.0 - use `autoPageRepeatHeader`
-	 */
-	addHeaderToEach?: boolean
-	/**
-	 * @deprecated v3.3.0 - use `autoPageSlideStartY`
-	 */
-	newSlideStartY?: number
 }
 export interface TableCellProps extends TextBaseProps {
 	/**
@@ -1575,11 +1478,6 @@ export interface TableProps extends PositionProps, TextBaseProps, ObjectNameProp
 	 * @default false // obviously
 	 */
 	verbose?: boolean // Undocumented; shows verbose output
-
-	/**
-	 * @deprecated v3.3.0 - use `autoPageSlideStartY`
-	 */
-	newSlideStartY?: number
 }
 export interface TableCell {
 	_type: SLIDE_OBJECT_TYPES.tablecell
@@ -1772,39 +1670,6 @@ export interface TextPropsOptions extends PositionProps, DataOrPathProps, TextBa
 	 * @default true
 	 */
 	wrap?: boolean
-
-	/**
-	 * Whether "Fit to Shape?" is enabled
-	 * @deprecated v3.3.0 - use `fit`
-	 */
-	autoFit?: boolean
-	/**
-	 * Whather "Shrink Text on Overflow?" is enabled
-	 * @deprecated v3.3.0 - use `fit`
-	 */
-	shrinkText?: boolean
-	/**
-	 * Inset
-	 * @deprecated v3.10.0 - use `margin`
-	 */
-	inset?: number
-	/**
-	 * Dash type
-	 * @deprecated v3.3.0 - use `line.dashType`
-	 */
-	lineDash?: 'solid' | 'dash' | 'dashDot' | 'lgDash' | 'lgDashDot' | 'lgDashDotDot' | 'sysDash' | 'sysDot'
-	/**
-	 * @deprecated v3.3.0 - use `line.beginArrowType`
-	 */
-	lineHead?: 'none' | 'arrow' | 'diamond' | 'oval' | 'stealth' | 'triangle'
-	/**
-	 * @deprecated v3.3.0 - use `line.width`
-	 */
-	lineSize?: number
-	/**
-	 * @deprecated v3.3.0 - use `line.endArrowType`
-	 */
-	lineTail?: 'none' | 'arrow' | 'diamond' | 'oval' | 'stealth' | 'triangle'
 }
 export interface TextProps {
 	text?: string
@@ -1988,15 +1853,6 @@ export interface IChartPropsBase {
 	 * @since v3.11
 	 */
 	plotArea?: IChartPropsFillLine
-
-	/**
-	 * @deprecated v3.11.0 - use `plotArea.border`
-	 */
-	border?: BorderProps
-	/**
-	 * @deprecated v3.11.0 - use `plotArea.fill`
-	 */
-	fill?: HexColor
 }
 export interface IChartPropsAxisCat {
 	/**
@@ -2467,11 +2323,6 @@ export interface SlideMasterProps {
 			text?: string
 		}
 	}>
-
-	/**
-	 * @deprecated v3.3.0 - use `background`
-	 */
-	bkgd?: string | BackgroundProps
 }
 export interface ObjectOptions extends ImageProps, PositionProps, ShapeProps, TableCellProps, TextPropsOptions {
 	_placeholderIdx?: number
@@ -2534,10 +2385,6 @@ export interface SlideBaseProps {
 	}>
 
 	background?: BackgroundProps
-	/**
-	 * @deprecated v3.3.0 - use `background`
-	 */
-	bkgd?: string | BackgroundProps
 }
 export type TransitionType = 'none' | 'fade' | 'push' | 'wipe' | 'cover' | 'split' | 'cut' | 'morph'
 export type TransitionDirection = 'left' | 'right' | 'up' | 'down'